fix: change request native permission as sub action
Changes over the whole logic of native permission's request to make it as a sub action and make the parent action more generic to, in the future, ask for different permission if needed
