+++ lib/agent/permissions/index.js
@@ -2,6 +2,7 @@ const osName = process.platform.replace('win32', 'windows');
 // eslint-disable-next-line import/no-dynamic-require
 const osFunctions = require(`./${osName}`);
 
-const { getLocationPermission } = osFunctions;
+const { getLocationPermission, requestNativeLocation } = osFunctions;
 
 exports.getLocationPermission = getLocationPermission;
+exports.requestNativeLocation = requestNativeLocation;
