+++ lib/agent/permissions/windows.js
@@ -1,4 +1,9 @@
 const getLocationPermission = () => {
 };
+const requestNativeLocation = (cb) => {
+  if (!cb || typeof cb !== 'function') return;
+  cb();
+};
 
 exports.getLocationPermission = getLocationPermission;
+exports.requestNativeLocation = requestNativeLocation;
